Financial Document Analyzer (CrewAI Multi-Agent System)/requirements.txt
click==8.1.7

# Required version (assignment instruction)
crewai==0.130.0
crewai-tools==0.47.1

fastapi==0.110.3
uvicorn==0.30.1

python-dotenv==1.0.1
python-multipart==0.0.9

langchain-core==0.1.52
langchain-community==0.0.38
langsmith==0.1.67

openai==1.30.5

pydantic==1.10.13
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0

pypdf==4.2.0
Jinja2==3.1.4
jsonschema==4.22.0
protobuf==4.25.3
oauthlib==3.2.2